Fix dead link in documentation.
This commit fixes the link to the form to sign up for the City Bureau Slack channel and resolves issue #155. 

I did fix a similar problem in `README.md` recently but didn't notice that we had a second reference to the same form in `docs/index.md`. In any case this was a nice test of the new dead link checker.