146: Update AIM Spider
This commit resolves an issue where our original solution fell out of sync
with the new domain and structure of the Authority for Improvements in
Municipalities meeting site.

Note that we did lose the ability to retrieve and parse the agendas and
minutes, but I don't consider that a priority at the moment. This
functionality could easily be added at a later date.