Add City of Pittsburgh Virtual Zoning Board
This commit adds a spider, tests, and a frozen HTML page for the new City of
Pittsburgh Virtual Zoning Board. This completes the replacement of the
"pitt_city_planning" spider.

Note that this spider is very similar to pitt_virtual_planning_commission as
the layout of the target websites are identical. This could have been an
opportunity to make a pitt_city_planning parent class with two child classes.
However, I decided against this because it would make the project structure
more complex and because when we have to fix a spider, we usually end up just
writing a new implementation from scratch because the structure of the target
page has changed substantially.