Add new and modify existing documentation
This commit moves material on general advice for building a spider from the "03-team.md" document to a new "Tips and Tricks" Markdown document, and expands on some of that information, pursuant with issue "Improve Tools Documentation #104".

The new "Tips and Tricks" document also adds paragraphs aimed to introduce novice programmers to the concepts of "Regular Expressions", "The Internet Wayback Machine", and "IDE" sections.